+++ vue/src/components/App.vue
@@ -1,26 +1,8 @@
 <template>
   <div class="app">
-    <nav>
-      <RouterLink :to="{ name: ROUTES.CATALOG }"> Каталог </RouterLink> |
-      <RouterLink :to="{ name: ROUTES.FAVORITES }"> Избранное </RouterLink> |
-      <RouterLink :to="{ name: ROUTES.CART }"> Корзина </RouterLink> |
-    </nav>
     <RouterView />
   </div> 
 </template>
-
-<script>
-  import { RouteNames } from '../router/routes'
-
-  export default {
-    computed: {
-      ROUTES () {
-        return RouteNames
-      }
-    }
-  }
-
-</script>
 
 <style lang="less">
 @import url("https://fonts.googleapis.com/css2?family=Jost:wght@400;700&display=swap");
