+++ vue/src/components/pages/Item.vue
@@ -21,34 +21,12 @@
                                 {{ item.price.currentPrice }}
                             </span>
                             <span class="item-page__current-price-hint">
-                                Цена по скидочной картой
+                                Цена по скидочной карте
                             </span>
                         </div>
                     </div>
-                    <CustomButton
-                        class="item-page__add-to-cart"
-                        type="filledOrange"
-                        @click="handleAddToCartClick"
-                        v-if="!cartItem"
-                    >
-                        В корзину
-                    </CustomButton>
-                    <div class="item-page__count__amount" v-if="cartItem">
-                        <button
-                            class="item-page__count__amount__button"
-                            @click="() => onChangeCount(-1)"
-                        >
-                            -
-                        </button>
-                        <div class="item-page__count__amount__number">
-                            {{ cartItem.selectCount }}
-                        </div>
-                        <button
-                            class="item-page__count__amount__button"
-                            @click="() => onChangeCount(1)"
-                        >
-                            +
-                        </button>
+                    <div class="item-page__add-to-cart">
+                        <ButtonCart :item=this.item />
                     </div>
                     <p class="item-page__description">
                         {{ item.description }}
@@ -76,26 +54,22 @@
 <script>
 import PageLayout from '../parts/PageLayout';
 import ImageSlider from '../parts/ImageSlider';
-import CustomButton from '../parts/Button.vue';
-import { mapGetters, mapActions } from 'vuex';
+import ButtonCart from '../parts/ButtonCart.vue';
+import { mapGetters } from 'vuex';
 
 export default {
     name: 'ItemPage',
     components: {
         PageLayout,
         ImageSlider,
-        CustomButton,
+        ButtonCart
     },
     computed: {
         ...mapGetters('catalog', ['getItemFromCatalogById']),
-        ...mapGetters('cart', ['getCart']),
         item() {
             return this.getItemFromCatalogById(
                 this.$route.params.itemId.slice(1)
             );
-        },
-        cartItem() {
-            return this.getCart.find((elem) => elem.id === this.item.id);
         },
         showPricesChange() {
             return (
@@ -104,38 +78,6 @@ export default {
         },
         getImage() {
             return this.item.images.find((image) => image.type === 'main').url;
-        },
-    },
-    methods: {
-        ...mapActions('cart', [
-            'addItemToCart',
-            'removeAllItemsFromCart',
-            'removeItemFromCart',
-            'changeItem',
-        ]),
-        handleAddToCartClick() {
-            this.addItemToCart({
-                id: this.item.id,
-                name: this.item.name,
-                brand: this.item.information[0].value,
-                price: {
-                    currentPrice: this.item.price.currentPrice,
-                    originalPrice: this.item.price.originalPrice,
-                },
-                images: this.item.images,
-                count: this.item.count,
-                selectCount: '1',
-            });
-        },
-        onChangeCount(value) {
-            const count = Number(this.cartItem.selectCount) + Number(value);
-            console.log(count, this.cartItem.count, count <= this.cartItem.count);
-            
-            if (count <= this.cartItem.count && count > 0) {
-                this.cartItem.selectCount = count;
-                this.changeItem(this.cartItem);
-            }
-            if (count < 1) this.removeItemFromCart(this.cartItem);
         },
     },
 };
@@ -174,32 +116,6 @@ export default {
         gap: 16px;
     }
 
-    &__add-to-cart {
-        width: 100%;
-        height: 60px;
-    }
-
-    &__add-to-cart-child {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        padding: 0 16px;
-    }
-
-    &__cart-btn-text {
-        height: 28px;
-        flex-grow: 1;
-
-        font-family: @ffOne;
-        font-size: 24px;
-        font-weight: 400;
-        text-align: center;
-    }
-
-    &__cart-btn-icon {
-        width: 32px;
-        height: 32px;
-    }
 
     &__prices-cnt {
         width: 100%;
@@ -246,6 +162,11 @@ export default {
         content: ' ₽';
     }
 
+    &__add-to-cart {
+        width: 100%;
+        height: 60px;
+    }
+
     &__description {
         font-family: @ffOne;
         font-size: 16px;
@@ -290,41 +211,6 @@ export default {
         font-weight: 700;
         line-height: 18px;
         text-align: left;
-    }
-
-    &__count {
-        display: flex;
-        align-items: center;
-        gap: 16px;
-        margin-left: auto;
-
-        &__amount {
-            display: flex;
-            align-items: center;
-            gap: 8px;
-            height: 40px;
-            padding: 8px;
-            background-color: @cBaseEleven;
-            border-radius: 4px;
-
-            &__button {
-                width: 24px;
-                height: 24px;
-                padding: 0;
-                background-color: @cBaseEleven;
-                color: @cBaseWhite;
-                border: none;
-                cursor: pointer;
-                font-size: 20px;
-                text-align: center;
-            }
-
-            &__number {
-                background-color: @cBaseEleven;
-                color: @cBaseWhite;
-                border: none;
-            }
-        }
     }
 }
 </style>
