+++ vue/src/components/parts/PageLayout.vue
@@ -1,14 +1,19 @@
 <template>
     <div>
         <div class="content">
+            <HeaderMenu> </HeaderMenu>
             <slot />
         </div>
     </div>
 </template>
 
 <script>
+import HeaderMenu from '../parts/Header';
 export default {
     name: 'PageLayout',
+    components: {
+        HeaderMenu,
+    },
 };
 </script>
